backend/pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "FastAPI backend for Study Buddy — document management, RAG pipeline, and user authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alabaster==1.0.0",
    "alembic==1.17.2",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "asyncpg==0.30.0",
    "attrs==25.4.0",
    "babel==2.17.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "build==1.3.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.2.0",
    "click==8.3.0",
    "coloredlogs==15.0.1",
    "distro==1.9.0",
    "docutils==0.22.3",
    "durationpy==0.10",
    "fastapi==0.119.0",
    "filelock==3.20.0",
    "flatbuffers==25.9.23",
    "fsspec==2025.9.0",
    "google-auth==2.41.1",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "grpcio==1.75.1",
    "h11==0.16.0",
    "hf-xet==1.1.10",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.35.3",
    "humanfriendly==10.0",
    "idna==3.11",
    "imagesize==1.4.1",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==34.1.0",
    "lxml==6.0.2",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "minio==7.2.20",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "networkx==3.6",
    "numpy==2.3.4",
    "numpydoc==1.10.0",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.5",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvshmem-cu12==3.3.20",
    "nvidia-nvtx-cu12==12.8.90",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.1",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.3",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "passlib[bcrypt]==1.7.4",
    "pillow==12.0.0",
    "polars==1.35.2",
    "polars-runtime-32==1.35.2",
    "posthog==5.4.0",
    "protobuf==6.33.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pycparser==2.23",
    "pycryptodome==3.23.0",
    "pydantic-core==2.41.4",
    "pydantic-settings>=2.14.0",
    "pydantic[email]==2.12.3",
    "pygments==2.19.2",
    "pyjwt>=2.12.1",
    "pypdf==6.4.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "python-magic==0.4.27",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis>=7.4.0",
    "referencing==0.37.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rich==14.2.0",
    "roman-numerals==3.1.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "slowapi>=0.1.9",
    "sniffio==1.3.1",
    "snowballstemmer==3.0.1",
    "sphinx==9.0.4",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sqlalchemy==2.0.44",
    "starlette==0.48.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.1",
    "torch==2.9.1",
    "torchaudio==2.9.1",
    "torchvision==0.24.1",
    "tqdm==4.67.1",
    "triton==3.5.1",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "zipp==3.23.0",
]
